refactor(web/applications): pr tweak, refactor into service file (BOAS-507)
